requirements.txt
Flask
Flask-CORS
firebase-admin
google-generativeai
numpy
scikit-learn
sentence-transformers
